segregate domain events and outbox
