+++ settings.gradle
@@ -1,5 +1,7 @@
 include ':xm-commons-config'
 include ':xm-commons-domain-event'
+include ':xm-commons-domain-event-db'
+include ':xm-commons-domain-event-outbox'
 include ':xm-commons-exceptions'
 include ':xm-commons-i18n'
 include ':xm-commons-lep'
