+++ xm-commons-domain-event-outbox/src/main/java/com/icthh/xm/commons/domain/event/outbox/domain/RecordStatus.java
@@ -1,0 +1,8 @@
+package com.icthh.xm.commons.domain.event.outbox.domain;
+
+public enum RecordStatus {
+    NEW,
+    PROCESSING,
+    COMPLETE,
+    ERROR
+}
