+++ xm-commons-domain-event-outbox/src/main/java/com/icthh/xm/commons/domain/event/outbox/domain/ValidFor.java
@@ -1,0 +1,19 @@
+package com.icthh.xm.commons.domain.event.outbox.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import javax.persistence.Embeddable;
+import java.time.Instant;
+
+@Embeddable
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ValidFor {
+    private Instant validFrom;
+    private Instant validTo;
+}
