+++ xm-commons-domain-event-outbox/src/main/resources/config/liquibase/changelog/outbox/00000000000001_initial_indexes.yaml
