+++ xm-commons-domain-event-outbox/src/test/resources/enabledDomainEvents.yml
@@ -1,0 +1,8 @@
+enabled: true
+sources:
+    DB:
+        enabled: true
+        transport: outboxTransport
+    WEB:
+        enabled: false
+        transport: outboxTransport
