+++ xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/domain/HttpDomainEventPayload.java
@@ -1,0 +1,22 @@
+package com.icthh.xm.commons.domain.event.domain;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+import java.net.http.HttpHeaders;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class HttpDomainEventPayload extends DomainEventPayload {
+    private String method;
+    private String url;
+    private String queryString;
+    private Long requestLength;
+    private String requestBody;
+    private Long responseLength;
+    private String responseBody;
+    private HttpHeaders requestHeaders;
+    private HttpHeaders responseHeaders;
+    private Integer responseCode;
+    private Long execTime;
+}
