--- xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/domain/RecordStatus.java
@@ -1,8 +1,0 @@
-package com.icthh.xm.commons.domain.event.domain;
-
-public enum RecordStatus {
-    NEW,
-    PROCESSING,
-    COMPLETE,
-    ERROR
-}
