+++ xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/service/Transport.java
@@ -1,6 +1,6 @@
 package com.icthh.xm.commons.domain.event.service;
 
-import com.icthh.xm.commons.domain.event.service.dto.DomainEvent;
+import com.icthh.xm.commons.domain.event.domain.DomainEvent;
 
 public interface Transport {
     void send(DomainEvent event);
