+++ xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/service/builder/DomainEventBuilder.java
@@ -1,6 +1,6 @@
 package com.icthh.xm.commons.domain.event.service.builder;
 
-import com.icthh.xm.commons.domain.event.service.dto.DomainEvent;
+import com.icthh.xm.commons.domain.event.domain.DomainEvent;
 
 public interface DomainEventBuilder {
     DomainEvent.DomainEventBuilder getPrefilledBuilder();
