+++ xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/service/builder/impl/DefaultDomainEventBuilder.java
@@ -1,7 +1,7 @@
 package com.icthh.xm.commons.domain.event.service.builder.impl;
 
 import com.icthh.xm.commons.domain.event.service.builder.DomainEventBuilder;
-import com.icthh.xm.commons.domain.event.service.dto.DomainEvent;
+import com.icthh.xm.commons.domain.event.domain.DomainEvent;
 import com.icthh.xm.commons.security.XmAuthenticationContext;
 import com.icthh.xm.commons.security.XmAuthenticationContextHolder;
 import com.icthh.xm.commons.tenant.TenantContextHolder;
@@ -13,7 +13,7 @@ import java.time.Instant;
 
 @Component("defaultDomainEventBuilder")
 @RequiredArgsConstructor
-class DefaultDomainEventBuilder implements DomainEventBuilder {
+public class DefaultDomainEventBuilder implements DomainEventBuilder {
 
     private final TenantContextHolder tenantContextHolder;
     private final XmAuthenticationContextHolder xmAuthenticationContextHolder;
