--- xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/service/dto/DomainEventPayload.java
@@ -1,23 +1,0 @@
-package com.icthh.xm.commons.domain.event.service.dto;
-
-import com.fasterxml.jackson.annotation.JsonSubTypes;
-import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.Map;
-
-@Data
-@NoArgsConstructor
-@JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "@type")
-@JsonSubTypes({
-    @JsonSubTypes.Type(value = DbDomainEventPayload.class, name = "DbDomainEventPayload"),
-    @JsonSubTypes.Type(value = HttpDomainEventPayload.class, name = "HttpDomainEventPayload")
-})
-public class DomainEventPayload {
-    private Map<String, Object> data;
-
-    public DomainEventPayload(Map<String, Object> data) {
-        this.data = data;
-    }
-}
