--- xm-commons-domain-event/src/main/java/com/icthh/xm/commons/domain/event/service/dto/ValidForDto.java
@@ -1,11 +1,0 @@
-package com.icthh.xm.commons.domain.event.service.dto;
-
-import lombok.Data;
-
-import java.time.Instant;
-
-@Data
-public class ValidForDto {
-    private Instant validFrom;
-    private Instant validTo;
-}
