+++ xm-commons-migration-db/build.gradle
@@ -1,3 +1,7 @@
+plugins {
+    id "maven-publish"
+}
+
 apply plugin: 'io.spring.dependency-management'
 dependencyManagement {
     imports {
@@ -19,4 +23,12 @@ dependencies {
 
     testCompile 'junit:junit'
     testCompile 'org.mockito:mockito-core'
+}
+
+publishing {
+    publications {
+        mavenJava(MavenPublication) {
+            from components.java
+        }
+    }
 }
